backend added

+++ backend/routes/posts.js
@@ -65,9 +65,13 @@ router.put("/:id",
       imagePath: imagePath
     })
     console.log(post)
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      console.log(result)
-      res.status(200).json({ message: "Updated successful" })
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Updated successful" })
+      } else {
+        res.status(401).json({ message: "Not Authorized" })
+      }
+
     })
   })
 
@@ -104,9 +108,12 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result)
-    res.status(200).json({ message: "Post Deleted!" })
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful" })
+    } else {
+      res.status(401).json({ message: "Not Authorized" })
+    }
   })
 
 })
